build: update L2AssetHandler tests to use uint256 risks

+++ test/facets/L2/AssetHandler/behaviors/withdrawERC721Assets.t.sol
@@ -206,8 +206,8 @@ contract L2AssetHandler_withdrawERC721Assets is
             )
         );
 
-        uint64 depositorTokenRisk = uint64(
-            uint256(vm.load(address(this), depositorTokenRiskStorageSlot))
+        uint256 depositorTokenRisk = uint256(
+            vm.load(address(this), depositorTokenRiskStorageSlot)
         );
 
         // this assertion proves that the risk for the depositor and the token ID in the collection was reset correctly
@@ -226,8 +226,8 @@ contract L2AssetHandler_withdrawERC721Assets is
             )
         );
 
-        uint64 tokenRisk = uint64(
-            uint256(vm.load(address(this), tokenRiskStorageSlot))
+        uint256 tokenRisk = uint256(
+            vm.load(address(this), tokenRiskStorageSlot)
         );
 
         // this assertion proves that the risk for the token ID in the collection was decremented correctly
@@ -261,8 +261,8 @@ contract L2AssetHandler_withdrawERC721Assets is
             )
         );
 
-        uint64 totalDepositorRisk = uint64(
-            uint256(vm.load(address(this), totalDepositorRiskStorageSlot))
+        uint256 totalDepositorRisk = uint256(
+            vm.load(address(this), totalDepositorRiskStorageSlot)
         );
 
         // this assertion proves that the total risk for the depositor in the collection was decremented correctly
@@ -276,8 +276,8 @@ contract L2AssetHandler_withdrawERC721Assets is
             )
         );
 
-        uint64 totalRisk = uint64(
-            uint256(vm.load(address(this), totalRiskStorageSlot))
+        uint256 totalRisk = uint256(
+            vm.load(address(this), totalRiskStorageSlot)
         );
 
         // this assertion proves that the total risk in the collection was decremented correctly
